refactor(fractal): split fractal-generator.ts into pipeline-stage modules (closes #281)
Move the 1241-line src/puzzle/fractal-generator.ts into src/puzzle/fractal/
along its natural pipeline boundaries:

- types.ts        — Tile, DiagonalConnection, ArcData
- tile.ts         — tileEq, makeTile, makeConnection, connectionFromQuad,
                    connectionKey
- cell-grid.ts    — CellGrid class
- arcs.ts         — makeArc, addArcs
- piece-growth.ts — createPiece, fillHoles, adoptOrphanTiles,
                    fillEmptyCells (+ findPossibleConnections,
                    hasPinwheelShape helpers)
- convert.ts      — convertToStandardPieces (+ borderPathBetween)
- index.ts        — public API (generateFractalPuzzle, scaleFractalGrid,
                    FractalConfig) plus the findDiagonalOwner helper that
                    only the entry point uses

The existing public-API tests follow the entry point to fractal/index.test.ts.
Importers in src/puzzle/index.ts, src/game/init.ts,
src/game/cut-style-strategies.ts, and scripts/diag-fractal-borders.ts now
point at './fractal/index.js'. No behaviour changes; all 1101 tests pass.
