+++ src/puzzle/index.ts
@@ -1,5 +1,5 @@
 export { generateProceduralPuzzle } from './procedural-generator.js';
-export { generateFractalPuzzle, scaleFractalGrid } from './fractal-generator.js';
+export { generateFractalPuzzle, scaleFractalGrid } from './fractal/index.js';
 export { generateComposablePuzzle } from './composable-generator.js';
 export type { ComposableConfig } from './composable-generator.js';
 export { createSeededRandom, generateSeed } from './seeded-random.js';
